### 2021-10-22 #3844 @whedon
[REVIEW]: compareMCMCs: An R package for studying MCMC efficiency
**Submitting author:** @perrydv (<a href="http://orcid.org/0000-0002-8329-6796">Perry de Valpine</a>)
**Repository:** <a href="https://github.com/nimble-dev/compareMCMCs/" target ="_blank">https://github.com/nimble-dev/compareMCMCs/</a>
**Version:** v1.0.0
