### 2021-10-28 nimble-dev/compareMCMCs#13 @rowlandseymour
Possibly unrelated Stan code 
In a couple of places there is some Stan code in the package. I'm not sure what the purpose of the code is, so I think it would be good to clarify or remove it.

The locations are:
https://github.com/nimble-dev/compareMCMCs/blob/7faa8c1cb251a5a20e4f5c744dd73b61ddf955ac/compareMCMCs/schools.stan#L1
https://github.com/nimble-dev/compareMCMCs/blob/7faa8c1cb251a5a20e4f5c744dd73b61ddf955ac/compareMCMCs/tests/testthat/schools.stan#L1

openjournals/joss-reviews#3844
