### 2021-10-28 nimble-dev/compareMCMCs#14 @rowlandseymour
Empty test file
The test file for the Stan test is empty.

The permalink is
https://github.com/nimble-dev/compareMCMCs/blob/7faa8c1cb251a5a20e4f5c744dd73b61ddf955ac/compareMCMCs/tests/testthat/test-stan.R

openjournals/joss-reviews#3844
