### 2021-10-28 nimble-dev/compareMCMCs#15 @rowlandseymour
Controlling for time
Currently users are allowed to specify the number of iterations and then ESS per second is reported as a measure. Have you considered allowing users to specify the computational time the MCMC can run for? To me at least, it makes sense to control for time rather than iterations.

This might be outside the remit of the JOSS review, so please feel free to push back on this.

A link the to user specified iterations is here:
https://github.com/nimble-dev/compareMCMCs/blob/7faa8c1cb251a5a20e4f5c744dd73b61ddf955ac/compareMCMCs/R/compareMCMCs.R#L67

openjournals/joss-reviews#3844
