### 2021-10-28 nimble-dev/compareMCMCs#16 @rowlandseymour
Stan example in Vignette
This is a great vignette; it's clear, accessible and has good examples in it. But I would have like to have seen an example of how Stan can be used here. I got it to work eventually but it involved a bit of trial and error. I know that different parts of statistics use different computational platforms, so I think including Stan would help open up the user base somewhat.

The only place Stan is currently mentioned in the vignette is here:
https://github.com/nimble-dev/compareMCMCs/blob/7faa8c1cb251a5a20e4f5c744dd73b61ddf955ac/compareMCMCs/vignettes/compareMCMCs.Rmd#L32

openjournals/joss-reviews#3844
