### 2021-10-28 nimble-dev/compareMCMCs#17 @rowlandseymour
Comments on paper
I think the paper is well written and does a good job at setting out the purpose of the software and making a case for its use. In my view, it's a perfectly acceptable example of what a JOSS paper should be. There are a couple of places where I think the language could be tidied up:

* l6: I think stochastically is a tautology here and "used to simulate" would be more appropriate.
* l9: Surely any kind of Bayesian statistics can require complex MCMC algorithms, not just applied?
* l19-21: The sentence beginning "Because MCMC samplers..." is quite long and complicated. I couldn't work out what was meant by this sentence.

openjournals/joss-reviews#3844
