### 2021-10-28 nimble-dev/compareMCMCs#18 @rowlandseymour
Contributor guidelines
One requirement for JOSS is that you have contributor guidelines. I can't find any currently for this package. Here's an example of some from a recently accepted JOSS paper

https://github.com/ejikeugba/serp/blob/43049d3c8fd55f51952e5354ccaa4997383442d0/contributing.md

openjournals/joss-reviews#3844
