### 2021-10-28 nimble-dev/compareMCMCs#14 @salleuska
Thanks @rowlandseymour !

@perrydv We ended up putting stan tests in the "general" file `test-compareMCMCs.R`. So I guess we may just want to remove that `test-stan.R`. Just mentioning because at first I did not remember this!
