### 2021-11-04 nimble-dev/compareMCMCs#19 @tbrown122387
logging cpu/machine details
Hey, you all mention a [few things that are timed](https://htmlpreview.github.io/?https://github.com/nimble-dev/compareMCMCs/blob/master/compareMCMCs/doc/compareMCMCs.html#what-is-timed) to help compare mcmcs. Do you guys also store information about the computer running the code? These times are highly dependent on that.

openjournals/joss-reviews#3844
