### 2021-11-04 nimble-dev/compareMCMCs#20 @tbrown122387
vignette formatting typo
Looks like something is going on with the phrase

We can see that the parameter is now log_a rather than a. The use of backticks in "log(\a`)" ensures thata` is evaluated as a variable and is described more below.

This is in the vignette in the section "Converting parameters to match each other across different MCMCs."

openjournals/joss-reviews#3844
