### 2021-11-04 nimble-dev/compareMCMCs#21 @tbrown122387
vignette example doesn't run correctly without rjags
The first example in the vignette does not work on my machine. The error seems to be `MCMCdef_jags_impl(MCMCinfo, MCMCcontrol, monitorInfo, modelInfo)` and it's complaining that `run_jags: rjags package is required for 'jags' option.  jags will not be run.` Perhaps `rjags` has to be added as a dependency? If that's too extreme, maybe add a line in the vignette, or a commented out `install.packages()` call or something.

Edit: same issue with the second example in the vignette: `rjags package is required for 'jags' option`.

Also, I get a deprecation warning, and it's not immediately obvious on how to turn it off.
```
> make_MCMC_comparison_pages(res, modelName = 'example1')
Warning message:
`guides(<scale> = FALSE)` is deprecated. Please use `guides(<scale> = "none")` instead.`
```

Here's the full output, for posterity.

```
> library(compareMCMCs)
Loading required package: R6
Loading required package: nimble
nimble version 0.12.1 is loaded.
For more information on NIMBLE and a User Manual,
please visit https://R-nimble.org.

Attaching package: ‘nimble’

The following object is masked from ‘package:stats’:

    simulate

Loading required package: ggplot2
Loading required package: grid
Loading required package: reshape2
Loading required package: xtable

Attaching package: ‘compareMCMCs’

The following object is masked from ‘package:nimble’:

    compareMCMCs

> modelCode <- nimbleCode({
+   a ~ dunif(0, 100)
+   y ~ dgamma(a, 2)
+ })
> modelInfo <- list(
+   code = modelCode,
+   constants = list(y = 2),
+   inits = list(a = 1)
+ )
> configure_nimble_slice <- function(model) {
+   configureMCMC(model, onlySlice = TRUE)
+ }
> res <- compareMCMCs(modelInfo,
+                     MCMCs = c('jags', 'nimble', 'nimble_slice'),
+                     nimbleMCMCdefs =
+                       list(nimble_slice = 'configure_nimble_slice'),
+                     MCMCcontrol = list(inits = list(a = 1),
+                                        niter = 2000,
+                                        burnin = 100))
building nimble model...
Defining model
  [Note] Using 'y' (given within 'constants') as data.
Building model
Setting data and initial values
Running calculate on model
  [Note] Any error reports that follow may simply reflect missing values in model variables.
Checking model sizes and dimensions
Error in MCMCdef_jags_impl(MCMCinfo, MCMCcontrol, monitorInfo, modelInfo) :
  run_jags: rjags package is required for 'jags' option.  jags will not be run.

===== Monitors =====
thin = 1: a
===== Samplers =====
RW sampler (1)
  - a
===== Monitors =====
thin = 1: a
===== Samplers =====
slice sampler (1)
  - a
Compiling
  [Note] This may take a minute.
  [Note] Use 'showCompilerOutput = TRUE' to see C++ compilation details.
Compiling
  [Note] This may take a minute.
  [Note] Use 'showCompilerOutput = TRUE' to see C++ compilation details.
|-------------|-------------|-------------|-------------|
|-------------------------------------------------------|
|-------------|-------------|-------------|-------------|
|-------------------------------------------------------|
Warning message:
In compareMCMCs(modelInfo, MCMCs = c("jags", "nimble", "nimble_slice"),  :
  MCMC  jags  failed.

> make_MCMC_comparison_pages(res, modelName = 'example1')
Warning message:
`guides(<scale> = FALSE)` is deprecated. Please use `guides(<scale> = "none")` instead.
```

openjournals/joss-reviews#3844
