### 2021-11-04 nimble-dev/compareMCMCs#22 @tbrown122387
does nimble need to be a dependency?
This is more of a question:

What happens if we remove `nimble` as a dependency? Should it be? If this is a standalone project, it might be desirable to remove it. Also, it looks like there are certain variable names that are shared  (e.g. `compareMCMCs()`), so you'll need to be vigilant about using the `::` operator.

https://github.com/nimble-dev/compareMCMCs/blob/7faa8c1cb251a5a20e4f5c744dd73b61ddf955ac/compareMCMCs/DESCRIPTION#L17

openjournals/joss-reviews#3844
