### 2021-11-04 nimble-dev/compareMCMCs#23 @tbrown122387
registerMCMCengine documentation 
In the documentation for `registerMCMCengine`, it mentions "For an example, see `compareMCMCs:::MCMCdef_jags`.' That's just a wrapper for `MCMCdef_jags_impl`, which isn't `@export`ed.

openjournals/joss-reviews#3844
