### 2021-11-05 nimble-dev/compareMCMCs#26 @fabian-s
Best practices for R packages not consistently followed
I see multiple issues in terms of missing documentation and poor namespace management. Please make sure your package goes through `R CMD check` / `devtools::check` cleanly.

```
> devtools::check("compareMCMCs")
[....]

── R CMD check results ───────────────────────────────── compareMCMCs 0.5.0 ────
Duration: 1m 23.3s

❯ checking dependencies in R code ... WARNING
  '::' or ':::' imports not declared from:
    ‘coda’ ‘rjags’ ‘rstan’
  'loadNamespace' or 'requireNamespace' calls not declared from:
    ‘coda’ ‘rjags’ ‘rstan’
  'library' or 'require' calls to packages already attached by Depends:
    ‘nimble’ ‘reshape2’
    Please remove these calls from your code.
  Packages in Depends field not imported from:
    ‘R6’ ‘ggplot2’ ‘grid’ ‘nimble’ ‘reshape2’ ‘xtable’
    These packages need to be imported from (in the NAMESPACE file)
    for when this namespace is loaded but not attached.
  Unexported object imported by a ':::' call: ‘nimble:::removeIndexing’
    See the note in ?`:::` about the use of this operator.

❯ checking for missing documentation entries ... WARNING
  Undocumented code objects:
    ‘MCMCmetric_efficiency’ ‘getMetrics’ ‘getPageComponents’
    ‘registerMetrics’ ‘registerPageComponents’ ‘unregisterMetric’
    ‘unregisterPageComponents’
  All user-level objects in a package should have documentation entries.
  See chapter ‘Writing R documentation files’ in the ‘Writing R
  Extensions’ manual.

❯ checking Rd \usage sections ... WARNING
  Undocumented arguments in documentation object 'combineMetrics'
    ‘include_times’

  Undocumented arguments in documentation object 'compareMCMCs'
    ‘metricOptions’ ‘conversions’ ‘needRmodel’ ‘verbose’

  Undocumented arguments in documentation object 'metrics'
    ‘...’

  Undocumented arguments in documentation object 'modifyMetrics'
    ‘options’

  Functions with \usage entries need to have the appropriate \alias
  entries, and all their arguments documented.
  The \usage entries must correspond to syntactically valid R code.
  See chapter ‘Writing R documentation files’ in the ‘Writing R
  Extensions’ manual.

❯ checking package dependencies ... NOTE
  Depends: includes the non-default packages:
    'R6', 'nimble', 'ggplot2', 'grid', 'reshape2', 'xtable'
  Adding so many packages to the search path is excessive and importing
  selectively is preferable.

❯ checking DESCRIPTION meta-information ... NOTE
  License components which are templates and need '+ file LICENSE':
    BSD_3_clause
  Package listed in more than one of Depends, Imports, Suggests, Enhances:
    ‘nimble’
  A package should be listed in only one of these fields.

❯ checking top-level files ... NOTE
  Non-standard file/directory found at top level:
    ‘schools.stan’

❯ checking R code for possible problems ... NOTE
  MCMCdef_dummy_impl: no visible global function definition for ‘rnorm’
  MCMCdef_jags_impl: no visible global function definition for ‘update’
  MCMCdef_jags_impl: no visible binding for '<<-' assignment to
    ‘monitorNodesBUGS’
  MCMCdef_jags_impl: no visible binding for global variable
    ‘monitorNodesBUGS’
  MCMCmetric_CI95low: no visible binding for global variable ‘quantile’
  MCMCmetric_CI95upp: no visible binding for global variable ‘quantile’
  MCMCmetric_median: no visible binding for global variable ‘median’
  MCMCmetric_sd: no visible binding for global variable ‘sd’
  allParamEfficiencyComparisonComponent: no visible global function
    definition for ‘aggregate’
  compareMCMCs: no visible global function definition for ‘nimbleModel’
  make_MCMC_comparison_pages: no visible global function definition for
    ‘jpeg’
  make_MCMC_comparison_pages: no visible global function definition for
    ‘dev.off’
  runNIMBLE: no visible global function definition for ‘buildMCMC’
  runNIMBLE: no visible global function definition for ‘compileNimble’
    aggregate buildMCMC compileNimble dev.off jpeg median
    monitorNodesBUGS nimbleModel quantile rnorm sd update
  Consider adding
    importFrom("grDevices", "dev.off", "jpeg")
    importFrom("stats", "aggregate", "median", "quantile", "rnorm", "sd",
               "update")
  to your NAMESPACE file.
```
