### 2021-11-05 #3844 @fabian-s
(sorry about those automatic reminders...)

@perrydv both reviewers have gone through their checklists and defined some issues in your repo.

I would consider

nimble-dev/compareMCMCs#16
nimble-dev/compareMCMCs#17
nimble-dev/compareMCMCs#18

to be issues currently blocking acceptance as they are directly relevant for points on the JOSS criteria checklist.
Your package currently also violates many community standards for best practice R packages, see nimble-dev/compareMCMCs#26.

Please adress/discuss the issues opened by the reviewers and me in the respective issue threads.
