### 2021-11-05 nimble-dev/compareMCMCs#26 @fabian-s
additionally from `goodpractice::gp()`:

```
✖ use '<-' for assignment instead of '='. '<-' is the standard, and R users
    and developers are used it and it is easier to read your code for them if you use
    '<-'.

    R/make_MCMC_comparison_pages.R:244:7
    R/make_MCMC_comparison_pages.R:315:6
    R/make_MCMC_comparison_pages.R:324:6
    R/make_MCMC_comparison_pages.R:426:6
    R/make_MCMC_comparison_pages.R:434:6

  ✖ avoid long code lines, it is bad for readability. Also, many people prefer
    editor windows that are about 80 characters wide. Try make your lines shorter than
    80 characters

    R/compareMCMCs.R:9:1
    R/compareMCMCs.R:28:1
    R/compareMCMCs.R:31:1
    R/compareMCMCs.R:33:1
    R/compareMCMCs.R:47:1
    ... and 197 more lines

  ✖ avoid calling setwd(), it changes the global environment. If you need it,
    consider using on.exit() to restore the working directory.

    R/make_MCMC_comparison_pages.R:101:3
    R/make_MCMC_comparison_pages.R:102:11

  ✖ avoid the library() and require() functions, they change the global search
    path. If you need to use other packages, import them. If you need to load them
    explicitly, then consider loadNamespace() instead, or as a last resort, declare them
    as 'Depends' dependencies.

    R/metrics-addMetrics.R:51:7
    R/runNIMBLE.R:8:3
   ```
