### 2021-11-29 ace0287
Round 1 of updates for JOSS reviewer comments
