+++ compareMCMCs/R/MCMCdef_jags.R
@@ -31,33 +31,37 @@ MCMCdef_jags_impl <- function(MCMCinfo,
     
     burninTime <- 0
     if(MCMCcontrol$burnin > 0) {
-        burninTimeResult <- system.time(update(object = jags_mod,
-                                               n.iter = MCMCcontrol$burnin))
+        burninTimeResult <- system.time(
+            update(object = jags_mod,
+                   n.iter = MCMCcontrol$burnin))
         burninTime <- burninTimeResult[3]
     }
 
     postburninTimeResult <- system.time({
-      jags_out <- rjags::coda.samples(model=jags_mod,
-                                      variable.names=monitorInfo$monitorVars,
-                                      n.iter=MCMCcontrol$niter - MCMCcontrol$burnin,
-                                      thin=MCMCcontrol$thin)
+        jags_out <- rjags::coda.samples(
+                               model=jags_mod,
+                               variable.names=monitorInfo$monitorVars,
+                               n.iter=MCMCcontrol$niter - MCMCcontrol$burnin,
+                               thin=MCMCcontrol$thin)
     })
     postburninTime <- postburninTimeResult[3]
     
-    monitorNodesBUGS <<- gsub(' ', '', monitorInfo$monitors)
+    monitorNodesBUGS <- gsub(' ', '', monitorInfo$monitors)
     samplesArray <- jags_out[[1]][, monitorNodesBUGS, drop=FALSE]
     result <- MCMCresult$new(samples = samplesArray,
-                             times = list(setup_system.time = setupTimeResult,
-                                          burnin_system.time = burninTimeResult,
-                                          postburnin_system.time = postburninTime,
-                                          setup = setupTime,
-                                          sampling = postburninTime + burninTime,
-                                          burnin = burninTime,
-                                          postburnin = postburninTime))
+                             times = list(
+                                 setup_system.time = setupTimeResult,
+                                 burnin_system.time = burninTimeResult,
+                                 postburnin_system.time = postburninTime,
+                                 setup = setupTime,
+                                 sampling = postburninTime + burninTime,
+                                 burnin = burninTime,
+                                 postburnin = postburninTime))
     unlink(modelFileName)
     return(result)
   } else {
-    stop("run_jags: rjags package is required for 'jags' option.  jags will not be run.\n")
+      stop(paste0("run_jags: rjags package is required for 'jags' option. ",
+                  "jags will not be run.\n"))
     return(NULL)
   }
 }
