+++ compareMCMCs/R/metrics-addMetrics.R
@@ -2,26 +2,40 @@
 #' 
 #' Clear metrics or add metrics to MCMC results.
 #' 
-#' @param results an `MCMCresult` object or list of `MCMCresult` objects.
-#' @param byParameter `TRUE` or `FALSE`: whether to clear `byParameter` metrics
+#' @param results an `MCMCresult` object or list of `MCMCresult`
+#'     objects.
+#' @param byParameter `TRUE` or `FALSE`: whether to clear
+#'     `byParameter` metrics
 #' @param byMCMC `TRUE` or `FALSE`: whether to clear `byMCMC` metrics
-#' @param metrics character vector of metric names to add.  See \code{\link{metrics}}.
+#' @param metrics character vector of metric names to add.  See
+#'     \code{\link{metrics}}.
+#' @param options named list of options.  When calling a metric
+#'     function (e.g. `mean`), if there is a named element with that
+#'     name (e.g. "mean"), it will be passed as the second argument to
+#'     the metric function.
 #' 
-#' @details 
-#' These functions provide ways to manipulate the collection of metrics inside one or more 
-#' `MCMCresult` objects.
+#' @details These functions provide ways to manipulate the collection
+#'   of metrics inside one or more `MCMCresult` objects.
 #' 
-#' The \code{\link{MCMCresult}} class is fairly simple.  One can also modify contents of an
-#' `MCMCresult` object using class methods or direct manipulation of contents.
+#' The \code{\link{MCMCresult}} class is fairly simple.  One can also
+#' modify contents of an `MCMCresult` object using class methods or
+#' direct manipulation of contents.
 #' 
-#' Metrics are organized as "byParameter", when there is one result for each parameter (column)
-#' of MCMC output, and "byMCMC", when there is one result for an entire MCMC sample (across all parameters).
+#' Metrics are organized as "byParameter", when there is one result
+#' for each parameter (column) of MCMC output, and "byMCMC", when
+#' there is one result for an entire MCMC sample (across all
+#' parameters).
 #' 
 #' `clearMetrics` clears all metrics by parameter, by MCMC, or both.
 #' 
-#' `addMetrics` populates a set of metrics.  See package vignette for more information.
+#' `addMetrics` populates a set of metrics.  See package vignette for
+#' more information.
 #' 
 #' @name modifyMetrics
+#'
+#' @aliases clearMetrics addMetrics
+#'
+#' @seealso \link{combineMetrics}
 
 #' @rdname modifyMetrics
 #' @export
@@ -48,8 +62,6 @@ addMetrics <- function(results,
                                    'ESS',
                                    'efficiency'),
                        options = list()) {
-  if(!require(reshape2)) stop(paste0('Package reshape2 is required to add metrics. ',
-                                     'Please install reshape2.'))
   if(!is.list(results))
     results <- list(results)
   if(is.character(metrics))
@@ -61,10 +73,11 @@ addMetrics <- function(results,
       thisMetricName <- metrics[[iM]]
       thisMetric <- compareMCMCs_registered_metrics[[ thisMetricName ]]
       if(is.null(thisMetric))
-        warning(paste0('No metric ',
-                       'named \"',
-                       thisMetricName,
-                       '\" is registered.  Use registerMetrics to add new metrics.'))
+        warning(
+          paste0('No metric ',
+                 'named \"',
+                 thisMetricName,
+                 '\" is registered.  Use registerMetrics to add new metrics.'))
     } else {
       thisMetric <- metrics[[iM]]
       thisMetricName <- paste0("#", iM)
@@ -95,11 +108,18 @@ addMetrics <- function(results,
 #' This is useful for seeing results from multiple MCMC engines compactly.
[…]

[…]
