### 2021-11-29 nimble-dev/compareMCMCs#13 @perrydv
Yes @rowlandseymour that was an extraneous stan file.  Thanks for noting that.  I removed it.

openjournals/joss-reviews#3844
