### 2021-11-29 nimble-dev/compareMCMCs#14 @perrydv
Fixed.

openjournals/joss-reviews#3844
