### 2021-11-29 nimble-dev/compareMCMCs#15 @perrydv
Hi @rowlandseymour.  Thanks for this suggestion.  It's a good idea, but the basic input to MCMC packages seems to be number of iterations, so we will stick with that.  (Whether managed by number of iterations or by computational time, it is important to get sufficiently long chains to obtain good estimates of ESS.).
