### 2021-11-29 nimble-dev/compareMCMCs#19 @perrydv
Hi @tbrown122387.  That's a great idea.  Thanks.  I added an option called `sessionInfo`, default `TRUE`, that will record the output of `sessionInfo()` in the `MCMCresult` object.  At the moment nothing is done with this, but it is there for inspection.  I will add an example in the vignette when I revise that.  One could go further with machine and session information, but this seems like a basic first step.  They would allow a user to check the provenance of various runs they may be comparing to each other.
