### 2021-11-29 nimble-dev/compareMCMCs#22 @perrydv
@tbrown122387 I did clean up other entries in Depends vs Imports, but it looks like we do need `nimble` in Depends. It's related to how nimble does a lot of R code generation and evaluation.  For example, it makes on-the-fly function definitions to use in models and class definitions for models and other components.  This brings up issues to do with `nimble` more than `compareMCMCs` so I won't go further.  If possible in the future we will be able to move `nimble` from Depends to Imports.

The reason for the duplication of the name `compareMCMCs` across packages is that the idea started as a function in `nimble` and now is being separated into its own package here (and completely re-designed).  So `compareMCMCs` in `nimble` just tells users they need this new package.  Once this is released, we will eventually remove `compareMCMCs` from `nimble`.
