### 2021-11-29 nimble-dev/compareMCMCs#26 @perrydv
Thanks @fabian-s .  I have fixed up all of these best practices recommendations.  The only exception is the warning on `setwd`, which recommends pairing it with `on.exit` if really needed.  That is the situation here, and we did have it paired with `on.exit` to reset the directory already.  I think that one needs to stay, so one warning from `goodpractice::gp()` remains for that reason.
