### 2021-11-30 nimble-dev/compareMCMCs#26 @fabian-s
using branch `joss` at https://github.com/nimble-dev/compareMCMCs/commit/ace0287fa1decc3489fcc85c74768d8be8be0e67,
I got:

1. an error trying to rebuild the vignettes, that was easy to fix like this::

```
--- a/compareMCMCs/vignettes/compareMCMCs.Rmd
+++ b/compareMCMCs/vignettes/compareMCMCs.Rmd
@@ -22,9 +22,9 @@ knitr::opts_chunk$set(
   collapse = TRUE,
   comment = "#>"
 )
-this_system_has_rjags <- requireNamespace(rjags)
+this_system_has_rjags <- requireNamespace("rjags")
 if(!this_system_has_rjags) message("Portions of this vignette use package rjags.  That is not installed, so those portions will be skipped.")
-this_system_has_rstan <- requireNamespace(rstan)
+this_system_has_rstan <- requireNamespace("rstan")
 if(!this_system_has_rstan) message("Portions of this vignette use package rstan.  That is not installed, so those portions will be skipped.")
 ```

2. after this fix, R CMD check still complains about:
 ```
── R CMD check results
──────────────────────────── compareMCMCs 0.5.0 ────
Duration: 1m 18.7s

> checking dependencies in R code ... WARNING
  '::' or ':::' imports not declared from:
    ‘rjags’ ‘rstan’
  'loadNamespace' or 'requireNamespace' calls not declared from:
    ‘rjags’ ‘rstan’
  Package in Depends field not imported from: ‘nimble’
    These packages need to be imported from (in the NAMESPACE file)
    for when this namespace is loaded but not attached.

> checking for code/documentation mismatches ... WARNING
  Codoc mismatches from documentation object 'compareMCMCs':
  compareMCMCs
    Code: function(modelInfo = list(), MCMCcontrol = list(niter = 10000,
                   thin = 1, burnin = 2000), MCMCs = "nimble", monitors =
                   character(), nimbleMCMCdefs = list(), externalMCMCinfo
                   = list(), metrics = c("mean", "median", "sd",
                   "CI95_low", "CI95_upp", "efficiency_coda"),
                   metricOptions = list(), conversions = list(), seed =
                   NULL, needRmodel, verbose = TRUE, sessionInfo = TRUE)
    Docs: function(modelInfo = list(), MCMCcontrol = list(niter = 10000,
                   thin = 1, burnin = 2000), MCMCs = "nimble", monitors =
                   character(), nimbleMCMCdefs = list(), externalMCMCinfo
                   = list(), metrics = c("mean", "median", "sd",
                   "CI95_low", "CI95_upp", "efficiency_coda"),
                   metricOptions = list(), conversions = list(), seed =
                   NULL, needRmodel, verbose = TRUE, sessionInfo = TRUE,
                   gc = TRUE)
    Argument names in docs not in code:
      gc

> checking R code for possible problems ... NOTE
  MCMCdef_jags_impl: no visible global function definition for ‘update’
    update
  Consider adding
    importFrom("stats", "update")
  to your NAMESPACE file.

> checking for unstated dependencies in vignettes ... NOTE
  'library' or 'require' calls not declared from:
    ‘rjags’ ‘rstan’
```
please repair remaining issues with dependencies, imports and method definitions and make sure your vignettes' code is executable..
