### 2021-11-30 nimble-dev/compareMCMCs#15 @rowlandseymour
That makes sense. Thanks for considering anyway.
