### 2021-12-05 5bf5684
added definition for jags burninTimeResult when burnin=0
