### 2021-12-05 9bf61cb
removed lp__ column from Stan samples array
