### 2022-01-07 ec6da99
Big fixes, test coverage, contributor guidelines, other steps for JOSS.
