+++ paper.md
@@ -25,9 +25,9 @@ bibliography: paper.bib
 
 # Summary
 
-Markov chain Monte Carlo (MCMC) algorithms are used to stochastically simulate from complicated probability distributions.  MCMC is very widely used to implement Bayesian statistical analysis, where the distribution of interest (the "target distribution") is a posterior distribution of parameters given data.  In this context, the posterior is known only up to a constant, so only relative probabilities (or densities) can be easily calculated, which is sufficient for MCMC to work.   In applied Bayesian statistics, MCMC algorithms for large or complex statistical models and data sets are sometimes run for minutes, hours or days, making them an analysis bottleneck, so there is a premium on efficiency.  MCMC efficiency includes both computational speed and algorithmic mixing, which refers to how well the algorithm explores the posterior distribution from one iteration to the next.  Computational speed may comprise one or more steps such as algorithm setup, MCMC "burn-in" or "warm-up" phases, and MCMC execution or "sampling". 
+Markov chain Monte Carlo (MCMC) algorithms are used to simulate from complicated probability distributions.  MCMC is very widely used to implement Bayesian statistical analysis, where the distribution of interest (the "target distribution") is a posterior distribution of parameters given data.  In this context, the posterior is known only up to a constant, so only relative probabilities (or densities) can be easily calculated, which is sufficient for MCMC to work.   In Bayesian statistical analysis, MCMC algorithms for large or complex statistical models and data sets are sometimes run for minutes, hours or days, making them an analysis bottleneck, so there is a premium on efficiency.  MCMC efficiency includes both computational speed and algorithmic mixing, which refers to how well the algorithm explores the posterior distribution from one iteration to the next.  Computational speed may comprise one or more steps such as algorithm setup, MCMC "burn-in" or "warm-up" phases, and MCMC execution or "sampling". 
 
-There are many MCMC algorithms (also called "samplers") and software packages implementing them.   Because MCMC samplers can be mixed and matched, with different samplers for the same or different dimensions of a target distribution possibly combined into the same MCMC algorithm, there is an enormous space of MCMC algorithms.  Invention of new methods, comparisons among methods, and theoretical study of MCMC mixing are all important areas of active research.  Various software packages provide samplers such as Gibbs, adaptive random-walk Metropolis-Hastings, slice, Hamiltonian, multivariate ("block") or other variants of these, and others.  Different MCMC algorithms can yield efficiency that differs by orders of magnitude for a particular problem, with these variations in efficiency being problem-dependent. 
+There are many MCMC algorithms (also called "samplers") and software packages implementing them.  Because MCMC samplers can be validly combined (e.g., iterated in sequence), for example with different samplers for different dimensions of a target distribution, there is an enormous space of MCMC methods.  Invention of new methods, comparisons among methods, and theoretical study of MCMC mixing are all important areas of active research.  Various software packages provide samplers such as Gibbs, adaptive random-walk Metropolis-Hastings, slice, Hamiltonian, multivariate ("block") or other variants of these, and others.  Different MCMC algorithms can yield efficiencies that differ by orders of magnitude for a particular problem, with these variations in efficiency being problem-dependent. 
 
[…]
