+++ compareMCMCs/doc/compareMCMCs.R
@@ -2,51 +2,89 @@
 ## Note: special steps are needed to build this vignette.
 ## This is because it generates html output that it links to.
 ## 1. devtools::build_vignettes(clean = FALSE)
-## 2. Move or copy the five files with "example1" in their name
+## 2. Move or copy the ten files with "example1" or "example2" in their name
 ##    from /vignettes to /doc.  This is the directory where build_vignettes
 ##    will have copied compareMCMCs.[R, Rmd, html].
 ## 3. Remove outputs from /vignettes if desired.
+##
+## If rstan generates C++ output with an error compiling foo.c,
+## try re-installing rstan. Clues suggest that resulting updates to StanHeaders
+## seems to fix the problem.
 knitr::opts_chunk$set(
   echo = TRUE,
   collapse = TRUE,
   comment = "#>"
 )
-
-## -----------------------------------------------------------------------------
 library(compareMCMCs)
+this_system_has_rjags <- requireNamespace("rjags")
+if(!this_system_has_rjags) message("Portions of this vignette use package rjags.  That is not installed, so those portions will be skipped.")
+this_system_has_rstan <- requireNamespace("rstan")
+if(!this_system_has_rstan) message("Portions of this vignette use package rstan.  That is not installed, so those portions will be skipped.")
+
+## ---- message=FALSE-----------------------------------------------------------
+# This model code will be used for both nimble and JAGS
 modelCode <- nimbleCode({
   a ~ dunif(0, 100)
   y ~ dgamma(a, 2)
 })
 modelInfo <- list(
   code = modelCode,
-  constants = list(y = 2),
+  constants = list(y = 2), # data can be included in constants or data list
   inits = list(a = 1)
 )
+# Here is a custom MCMC configuration function for nimble
 configure_nimble_slice <- function(model) {
   configureMCMC(model, onlySlice = TRUE)
 }
+# Here is information to set the model up for Stan
+stan_code <- c("data {real y;}",
+               "parameters {real a;}",
+               "model {target += uniform_lpdf(a | 0, 100);",
+               "       target += gamma_lpdf(y | a, 2);}")
+writeLines(stan_code, "toy.stan")
+# An alternative way to provide the Stan model is to make a
+# list entered below as stan_model_args (similar to how sampling_args
+# appears in the externalMCMCinfo list).  The elements of stan_model_args
+# (e.g. 'file' or 'model_code') will be passed to the 'stan_model` function.
+stan_sampling_args <- list(data = list(y=2),
+                           init = list(list(a=1)), # Only one chain will be run
+                           warmup = 200, # Stan was not happy with warmup = 100
+                           iter = 2000)
+# Here is the call to compareMCMCs
 res <- compareMCMCs(modelInfo,
-                    MCMCs = c('jags', 'nimble', 'nimble_slice'),
+                    MCMCs = c(if(this_system_has_rjags) 'jags' else NULL,
+                              'nimble',       # nimble with default samplers
+                              'nimble_slice', # nimble with slice samplers
+                              if(this_system_has_rstan) 'stan' else NULL),
                     nimbleMCMCdefs = 
                       list(nimble_slice = 'configure_nimble_slice'),
                     MCMCcontrol = list(inits = list(a = 1),
                                        niter = 2000,
-                                       burnin = 100))
+                                       burnin = 100),
+                    externalMCMCinfo = list(stan = list(
+                      file = "toy.stan",
+                      sampling_args = stan_sampling_args))
+                    # Stan was not happy with warmup = 100, which would match the others.
+)
+file.remove("toy.stan")
 make_MCMC_comparison_pages(res, modelName = 'example1')
 
-## -----------------------------------------------------------------------------
[…]

[…]
