### 2022-01-07 nimble-dev/compareMCMCs#16 @perrydv
Hi @rowlandseymour Thanks for the suggestion.  We've added a Stan example to the vignette and also included Stan in test coverage.
