### 2022-01-07 nimble-dev/compareMCMCs#17 @perrydv
Hi @rowlandseymour Thanks for giving this a detailed read.  We've revised accordingly.
