### 2022-01-07 nimble-dev/compareMCMCs#18 @perrydv
Hi @rowlandseymour Thanks for catching that.  We added contributor guidelines at the end of the README.md.
