### 2022-01-07 nimble-dev/compareMCMCs#20 @perrydv
Thanks for catching that @tbrown122387.  This is fixed now.
