### 2022-01-07 nimble-dev/compareMCMCs#21 @perrydv
Hi @tbrown122387 I think this is fixed now. The vignette starts by checking if `rjags` and/or `rstan` are installed and only runs relevant code if they are.  The deprecrated feature was something in `ggplot2`.  We updated our code accordingly.
