### 2022-01-07 c83e743
tweak docs and import sd from stats
