+++ compareMCMCs/NAMESPACE
@@ -31,4 +31,5 @@ importFrom(nimble,buildMCMC)
 importFrom(nimble,compileNimble)
 importFrom(nimble,configureMCMC)
 importFrom(nimble,nimbleModel)
+importFrom(stats,sd)
 importFrom(stats,update)
