### 2022-01-07 nimble-dev/compareMCMCs#26 @perrydv
Thanks @fabian-s. It was in an intermediate state last time you prodded it.  For me it now passes R CMD check cleanly and the vignette builds cleanly.
