### 2022-01-07 #3844 @perrydv
@fabian-s Thanks for pinging me on this.  I think all issues are now handled with corresponding replies on the [github issues threads](https://github.com/nimble-dev/compareMCMCs/issues).
