### 2022-01-08 nimble-dev/compareMCMCs#21 @tbrown122387
Hey @perrydv I can get rid of `"jags"` from `MCMCs=`, but if I don't I get this when I re-run everything.

```
> library(compareMCMCs)
Loading required package: R6
Loading required package: nimble
nimble version 0.12.1 is loaded.
For more information on NIMBLE and a User Manual,
please visit https://R-nimble.org.

Attaching package: ‘nimble’

The following object is masked from ‘package:stats’:

    simulate

Loading required package: ggplot2
Loading required package: grid
Loading required package: reshape2
Loading required package: xtable

Attaching package: ‘compareMCMCs’

The following object is masked from ‘package:nimble’:

    compareMCMCs

> modelCode <- nimbleCode({
+   a ~ dunif(0, 100)
+   y ~ dgamma(a, 2)
+ })
> modelInfo <- list(
+   code = modelCode,
+   constants = list(y = 2),
+   inits = list(a = 1)
+ )
> configure_nimble_slice <- function(model) {
+   configureMCMC(model, onlySlice = TRUE)
+ }
> res <- compareMCMCs(modelInfo,
+                     MCMCs = c('jags', 'nimble', 'nimble_slice'),
+                     nimbleMCMCdefs =
+                       list(nimble_slice = 'configure_nimble_slice'),
+                     MCMCcontrol = list(inits = list(a = 1),
+                                        niter = 2000,
+                                        burnin = 100))
building nimble model...
Defining model
  [Note] Using 'y' (given within 'constants') as data.
Building model
Setting data and initial values
Running calculate on model
  [Note] Any error reports that follow may simply reflect missing values in model variables.
Checking model sizes and dimensions
Error in MCMCdef_jags_impl(MCMCinfo, MCMCcontrol, monitorInfo, modelInfo) :
  run_jags: rjags package is required for 'jags' option.  jags will not be run.

===== Monitors =====
thin = 1: a
===== Samplers =====
RW sampler (1)
  - a
===== Monitors =====
thin = 1: a
===== Samplers =====
slice sampler (1)
  - a
Compiling
  [Note] This may take a minute.
  [Note] Use 'showCompilerOutput = TRUE' to see C++ compilation details.
Compiling
  [Note] This may take a minute.
  [Note] Use 'showCompilerOutput = TRUE' to see C++ compilation details.
|-------------|-------------|-------------|-------------|
|-------------------------------------------------------|
|-------------|-------------|-------------|-------------|
|-------------------------------------------------------|
Warning message:
In compareMCMCs(modelInfo, MCMCs = c("jags", "nimble", "nimble_slice"),  :
  MCMC  jags  failed.

> make_MCMC_comparison_pages(res, modelName = 'example1')
Warning message:
`guides(<scale> = FALSE)` is deprecated. Please use `guides(<scale> = "none")` instead.
> res_jags <- compareMCMCs(modelInfo,
+                 MCMCs = c('jags'),
+                 MCMCcontrol = list(inits = list(a = 1),
+                                    niter = 2000,
+                                    burnin = 100))
building nimble model...
Defining model
  [Note] Using 'y' (given within 'constants') as data.
Building model
Setting data and initial values
Running calculate on model
  [Note] Any error reports that follow may simply reflect missing values in model variables.
Checking model sizes and dimensions
Error in MCMCdef_jags_impl(MCMCinfo, MCMCcontrol, monitorInfo, modelInfo) :
  run_jags: rjags package is required for 'jags' option.  jags will not be run.

Warning message:
In compareMCMCs(modelInfo, MCMCs = c("jags"), MCMCcontrol = list(inits = list(a = 1),  :
  MCMC  jags  failed.

>
```

It's kind of weird because I have `rjags` installed

```
> which(grepl("rjags", installed.packages()))
[1] 123
```

I figured it was because I installed `rjags` but not JAGS, but apparently that's not true. Both installed without any warnings or errors, but `library(rjags)` yields this:

```
> library(rjags)
Error: package or namespace load failed for ‘rjags’:
 .onLoad failed in loadNamespace() for 'rjags', details:
  call: dyn.load(file, DLLpath = DLLpath, ...)
  error: unable to load shared object '/Library/Frameworks/R.framework/Versions/4.1/Resources/library/rjags/libs/rjags.so':
  dlopen(/Library/Frameworks/R.framework/Versions/4.1/Resources/library/rjags/libs/rjags.so, 10): Library not loaded: /usr/local/lib/libjags.4.dylib
  Referenced from: /Library/Frameworks/R.framework/Versions/4.1/Resources/library/rjags/libs/rjags.so
  Reason: image not found
```

I don't even have the directory `/Library/Frameworks/` so this is strange.
