### 2022-01-08 nimble-dev/compareMCMCs#21 @perrydv
Hi @tbrown122387 Thanks for trying this.  I tried uninstalling rjags and I was still able to build the vignette cleanly without it.  Is the output you're showing from running the vignette code directly (i.e. not knitting it)?  If so, it looks like the errors are safe and accurate, saying that jags was not run.  When building the vignette, a variable `this_system_has_rjags` prevents some code chunks from running at all, to avoid those (valid) errors.  Also I no longer have the deprecation messages.  Did you reinstall `compareMCMCs` from the `joss` branch?  That's where I put the changes.  I'll go ahead and move them to `master`.  Finally I a wondering if the final error you're showing is entirely related to rjags and JAGS, not `compareMCMCs`?  The JAGS installation is outside of R.
