### 2022-01-09 nimble-dev/compareMCMCs#21 @tbrown122387
@perrydv yes you're right--my not having jags/rjags installed correctly on this machine has nothing to do with your work. The errors are indeed nonfatal/safe/best-case-scenario.

Regarding deprecation stuff, good call...that was fixed with `git pull`.
