### 2022-01-10 nimble-dev/compareMCMCs#16 @rowlandseymour
Thanks @perrydv. I think this really helps.
