### 2022-01-10 nimble-dev/compareMCMCs#26 @fabian-s
Awesome, looking great now!
