### 2022-01-10 #3844 @fabian-s
@perrydv

Almost done :partying_face:  -- at this point could you please:

- [ ] Add the missing DOIs for BUGS and MCMCpack to your references (see above)
