### 2022-01-11 #3844 @perrydv
Thanks @fabian-s.  Can you help me understand what is the best step in the process to put our package on CRAN?  After all of the above steps, so that we can then list the resulting JOSS publication as the CITATION request for the package?
