### 2022-01-12 #3844 @fabian-s
For that you'll need the DOI of the JOSS paper, which is already given on your draft?
Not sure I really understand the question, sorry -- having a JOSS publication and getting on CRAN seem to be unrelated to me.
