### 2022-01-12 #3844 @tbrown122387
@perrydv all the details for an R package’s`CITATION` file would be ready shortly after publishing with JOSS
