### 2022-01-12 cee9f1e
Fix vignette formatting detail.  Add some DOIs to paper.bib.
