+++ paper.bib
@@ -82,6 +82,7 @@
   number    = {25},
   pages     = {3049--3067},
   year      = {2009},
+  doi       = {10.1002/sim.3680},
   publisher = {Wiley Online Library}
 }
 
@@ -93,6 +94,7 @@
   number    = {4},
   pages     = {325--337},
   year      = {2000},
+  doi       = {10.1023/A:1008929526011},
   publisher = {Springer}
 }
 
