+++ compareMCMCs/doc/compareMCMCs.Rmd
@@ -203,8 +203,8 @@ Let's look at an example of each.  Say we want to make a metric that calculates 
 MCMCmetric_quartiles <- function(result, options) {
   p25 <- apply(result$samples, 2, quantile, probs = 0.25)
   p75 <- apply(result$samples, 2, quantile, probs = 0.75)
-  ## q25 and q75 are named vectors with names matching model parameters
-  ## i.e. column names of result$samples
+  # q25 and q75 are named vectors with names matching model parameters
+  # i.e. column names of result$samples
   maxDiff <- max(p75-p25)
   list(byParameter = list(p25 = p25,
                           p75 = p75),
@@ -254,8 +254,8 @@ res <- compareMCMCs(modelInfo,
                                        niter = 2000,
                                        burnin = 100))
 
-## We will look at the result using combineMetrics
-## rather than generating new html pages.
+# We will look at the result using combineMetrics
+# rather than generating new html pages.
 combineMetrics(res)
 ```
 
@@ -273,7 +273,7 @@ conversions <- list(nimble = reparam,
 applyConversions(res, conversions)
 clearMetrics(res)
 addMetrics(res) # use default metrics
-combineMetrics(res) ## An easy way to see that it worked
+combineMetrics(res) # An easy way to see that it worked
 ```
 
 ## How to specify conversions
