### 2022-01-12 0abdc54
Add DOI for MCMCpack
