+++ paper.bib
@@ -186,6 +186,7 @@ url     = {https://CRAN.R-project.org/view=Bayesian}
   volume    = {42},
   number    = {9},
   pages     = {22},
+  doi       = {10.18637/jss.v042.i09},
   url       = {https://www.jstatsoft.org/v42/i09/}
 }
 
