### 2022-01-13 #3844 @perrydv
Thanks @fabian-s.  I've done the following:

- Fixed a formatting detail in the vignette.
- Added three DOIs to paper.bib and re-ran whedon to build the pdf and check the DOIs.
- Tagged a release as v1.0.0 on GitHub.
- Archived the release on Zenodo.
- Found where to update author names in Zenodo metadata and title to match the paper title.
- The DOI from Zenodo is: 10.5281/zenodo.5842623

The process is new to me and I was not sure whether to take the (permanent) step of hitting publish on Zenodo.
