### 2022-01-13 #3844 @fabian-s
Thanks -- I see:
![image](https://user-images.githubusercontent.com/998541/149305500-546d6695-516d-47a5-a588-acc9a37d10a9.png)

-- we prefer that the software author names match those on the JOSS paper. Is there a reason you didn't use your real name instead of `perrydv`?
