### 2022-01-13 #3844 @perrydv
@fabian-s Zenodo did something where it automatically populated metadata and took my GitHub user name (perrydv) instead of my real name.  I thought I fixed that when I updated the metadata.  Also I thought I updated the title to match the paper title as requested.  I went back into that and tried save and publish.  Does this help: https://zenodo.org/record/5842623.  Sorry I'm new to this and not sure I've done it right.
