### 2022-01-14 #3844 @danielskatz
Congratulations to @perrydv (Perry de Valpine) and co-authors!!

And thanks to @rowlandseymour @tbrown122387 for reviewing, and to @fabian-s for editing!
We couldn't do this without your efforts!

(and thanks specifically to @perrydv and @fabian-s - it's a pleasure for me as AEiC to get a paper that's completely ready for publishing, with no additional editing needed)
